Continued refactor for saphyr/traction/tubes
Moved all things store for tubes into a submodule,
as it was getting overcrowded.
