+++ src/store/traction/saphyr/mutations.js
@@ -1,7 +1,4 @@
 const mutations = {
-  setTubes (state, tubes) {
-    state.tractionTubes = tubes
-  },
   setRuns (state, runs) {
     state.runs = runs
   },
