+++ src/store/traction/saphyr/tubes/getters.js
@@ -1,0 +1,7 @@
+const getters = {
+  tractionTubes:   state => state.tractionTubes,
+  tubeRequest:     (state, getters, rootState) => rootState.api.traction.saphyr.tubes,
+  requestsRequest: (state, getters, rootState) => rootState.api.traction.saphyr.requests,
+}
+
+export default getters
