+++ src/store/traction/saphyr/tubes/index.js
@@ -2,17 +2,13 @@ import state from './state'
 import getters from './getters'
 import mutations from './mutations'
 import actions from './actions'
-import saphyrTubesModule from '@/store/traction/saphyr/tubes/index'
 
-const saphyr = {
+const tubes = {
   namespaced: true,
   state,
   getters,
   mutations,
-  actions,
-  modules: {
-    tubes: saphyrTubesModule
-  }
+  actions
 }
 
-export default saphyr
+export default tubes
