+++ src/store/traction/saphyr/tubes/mutations.js
@@ -1,0 +1,7 @@
+const mutations = {
+  setTubes (state, tubes) {
+    state.tractionTubes = tubes
+  }
+}
+
+export default mutations
