+++ src/store/traction/saphyr/tubes/state.js
@@ -1,6 +1,5 @@
 const state = {
-  runs: [],
-  currentRun: {}
+  tractionTubes: []
 }
 
 export default state
